requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
PyPDF2==3.0.1
langchain==0.0.207
openai==0.27.6
python-dotenv==1.0.0
faiss-cpu==1.7.3